Finalize Profile and add example. (#5533)
This PR is intended to be the final Profile-related update before the
first release that includes the Profile feature. It includes the
following changes:
- Addresses the comment in
https://github.com/TileDB-Inc/TileDB/pull/5498#discussion_r2116503423
regarding the use of the `.tiledb` path extension: it is now only
appended to the default path (homedir). If the user provides a custom
path, nothing is appended.
- Removes the constructor of the `RestProfile` class that accepts only
the profile name, by assigning default values to the parameters of the
other constructor. This simplifies the overall logic.
- Makes the internal profile removal function, as well as the C++ API,
static. This better aligns with the intended use cases of the feature,
as Profiles are rarely created, and it is not logical to require users
to create a Profile object just to remove a profile from disk.
- Adds a C++ example demonstrating the expected usage of Profiles.

Closes CORE-235

---
TYPE: IMPROVEMENT
DESC: Finalize Profile feature and add example.

+++ scripts/run-nix-examples.sh
@@ -37,6 +37,10 @@ do
   if [ "${example##*/}" == png_ingestion_webp.cc ]; then
     continue
   fi;
+  # Skip Profile example as it requires serialization
+  if [ "${example##*/}" == profile.cc ]; then
+    continue
+  fi;
 
   cd ${TestAppDir}
   rm -rf ${TestAppDataDir}
